SQL code for database, hideous feed.php file for testing purposes
If you execute the news feed database.sql file in adminer, then upload the news folder here to your remote wn17 directory, the feeds should should use the information in your database to query Google and display the feeds. It's indescribably hideous, of course, but the SQL part seems to work.